+++ server/application-server/src/main/java/de/tum/in/www1/hephaestus/mentor/message/MessageRepository.java
@@ -1,4 +1,4 @@
-package de.tum.in.www1.hephaestus.chat.message;
+package de.tum.in.www1.hephaestus.mentor.message;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
